chore: address issue #610
